Generate json from tasks and team excel files
