+++ SpreadsheetUtility.UI.Web/Components/Layout/NavMenu.razor
@@ -31,7 +31,13 @@
                 <span class="bi bi-plus-square-fill-nav-menu" aria-hidden="true"></span> Gantt Generator
             </NavLink>
         </div>
-        
+
+        <div class="nav-item px-3">
+            <NavLink class="nav-link" href="ganttGeneratorFromPaste">
+                <span class="bi bi-plus-square-fill-nav-menu" aria-hidden="true"></span> Gantt Generator Paste
+            </NavLink>
+        </div>
+
         <div class="nav-item px-3">
             <NavLink class="nav-link" href="jsonGeneratorFromPaste">
                 <span class="bi bi-plus-square-fill-nav-menu" aria-hidden="true"></span> JSON Generator
