+++ SpreadsheetUtility.UI.Web/Components/Pages/JsonGeneratorFromPaste.razor
@@ -1,4 +1,4 @@
-﻿@page "/ganttGeneratorPaste"
+﻿@page "/jsonGeneratorFromPaste"
 @inject IGanttChartService GanttChartService
 @rendermode InteractiveServer
 @using Newtonsoft.Json
